Restructure packaging for online banking PL
COAND-497
